variable_task/requirements.txt
python-telegram-bot==20.7
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
matplotlib==3.8.2
seaborn==0.13.1
Pillow==10.2.0
APScheduler==3.10.4
python-dotenv==1.0.1 